Refactor scrolled window implementation and improve configuration handling by adding call back function
