+++ src/widgets/button/radio_button.c
@@ -85,68 +85,8 @@ ViewConfig *configure_radio_button_property(RadioButtonConfig *radio_button_conf
 
 ViewConfig *init_radio_button_config(FILE *index, RadioButtonConfig *radio_button_config)
 {
-    // Check if the radio_button config and the index file is not null
-    if (!radio_button_config || !index)
-        return NULL;
-
-    // Create view config
-    ViewConfig *view_config = NULL;
-    SAFE_ALLOC(view_config, ViewConfig, 1);
-    DFEAULT_VIEW_CONFIG(view_config);
-
-    // Store the property and value of the tag
-    gchar *property = NULL;
-    gchar *value = NULL;
-
-    // Read the tag character by character
-    gchar c;
-    while ((c = fgetc(index)) != '>')
-    {
-        /* If the character is a letter then go back one character
-            Because when the tag is readed the cursor will start with the first letter in the property and it will be lost */
-        if (is_character(c))
-            fseek(index, -1, SEEK_CUR);
-
-        int status = -1;
-
-        // Read the property of the tag
-        property = read_property(index, &status);
-
-        // If the all properties are readed then break the loop and return the view id and pass the properties to the radio_button config
-        if (status == 2)
-            return view_config;
-
-        // If the property is readed then read the value of the property
-        else if (status == 1 && property)
-        {
-            // Read the value of the property
-            value = read_value(index, &status);
-            if (status == 1 && value)
-            {
-                if (g_strcmp0(property, "id") == 0) // Store the view id
-                {
-                    strcpy(view_config->view_id, value);
-                    free(property);
-                }
-                else
-                {
-                    // Apply the property value to the radio_button config
-                    int config_status;
-                    view_config = configure_radio_button_property(radio_button_config, view_config, property, value, &config_status);
-                    if (!config_status)
-                    {
-                        printf("Error configuring the property %s\n", property);
-                    }
-                    free(value);
-                    free(property);
-                }
-            }
-        }
-    }
-
-    return view_config;
+    return init_generic_config(index,(void*)radio_button_config,(ConfigurePropertyCallback)configure_radio_button_property);
 }
-
 GtkWidget *create_radio_button(RadioButtonConfig radio_button_config)
 {
     GtkWidget *radio_button = NULL;
