+++ src/widgets/menu/menu_item.c
@@ -57,63 +57,8 @@ ViewConfig *configure_menu_item_property(MenuItemConfig *menu_item_config, ViewC
 
 ViewConfig *init_menu_item_config(FILE *index, MenuItemConfig *menu_item_config)
 {
-    // Check if the menu_item config and the index file is not null
-    if (!menu_item_config || !index)
-        return NULL;
-
-    // Create view config
-    ViewConfig *view_config = NULL;
-    SAFE_ALLOC(view_config, ViewConfig, 1);
-    DFEAULT_VIEW_CONFIG(view_config);
-
-    // Store the property and value of the tag
-    gchar *property = NULL;
-    gchar *value = NULL;
-
-    // Read the tag character by character
-    gchar c;
-    while ((c = fgetc(index)) != '>')
-    {
-        /* If the character is a letter then go back one character
-            Because when the tag is readed the cursor will start with the first letter in the property and it will be lost */
-        if (is_character(c))
-            fseek(index, -1, SEEK_CUR);
-
-        int status = -1;
-
-        // Read the property of the tag
-        property = read_property(index, &status);
-
-        // If the all properties are readed then break the loop and return the view id and pass the properties to the menu_item config
-        if (status == 2)
-            return view_config;
-
-        // If the property is readed then read the value of the property
-        else if (status == 1 && property)
-        {
-            // Read the value of the property
-            value = read_value(index, &status);
-            if (status == 1 && value)
-            {
-                if (g_strcmp0(property, "id") == 0) // Store the view id
-                {
-                    strcpy(view_config->view_id, value);
-                    free(property);
-                }
-                else
-                {
-                    // Apply the property value to the menu_item config
-                    view_config = configure_menu_item_property(menu_item_config, view_config, property, value);
-                    free(value);
-                    free(property);
-                }
-            }
-        }
-    }
-
-    return view_config;
+    return init_generic_config(index,(void*)menu_item_config,(ConfigurePropertyCallback)configure_menu_item_property);
 }
-
 GtkWidget *create_menu_item(MenuItemConfig menu_item_config)
 {
     GtkWidget *menu_item = NULL;
